Restructure and refactor Auth feature
