--- Frontend/src/features/auth/context/index.ts
@@ -1,2 +1,0 @@
-export * from './authContext';
-export * from './authProvider';
