--- Frontend/src/features/auth/css/Login.css
@@ -1,14 +1,0 @@
-#login-page {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-
-  padding-block: 2rem;
-}
-
-#login-page .login-form fieldset {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.5rem;
-  flex-direction: column;
-}
