--- Frontend/src/features/auth/css/styles.css
@@ -1,1 +1,0 @@
-@import url('./Login.css');
