--- Frontend/src/features/auth/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useAuthContext';
