--- Frontend/src/features/auth/loaders/index.ts
@@ -1,1 +1,0 @@
-export * from './requireAuthLoader';
