--- Frontend/src/features/shared/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useFetchWithToken';
