+++ Frontend/src/main.tsx
@@ -2,7 +2,7 @@ import './styles/index.css';
 
 import { createRoot } from 'react-dom/client';
 import { router } from './router';
-import { AuthProvider } from './features/auth/context/authProvider';
+import { AuthProvider } from './utilities/context/auth/authProvider';
 import { RouterProvider } from 'react-router';
 
 createRoot(document.getElementById('root')!).render(
