+++ Frontend/src/styles/global.css
@@ -9,3 +9,18 @@
 
   padding-block: 1rem;
 }
+
+#login-page {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  padding-block: 2rem;
+}
+
+#login-page .login-form fieldset {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.5rem;
+  flex-direction: column;
+}
