+++ Frontend/src/styles/index.css
@@ -1,4 +1,3 @@
 @import url('./base.css');
 @import url('./global.css');
-@import url('../features/auth/css/styles.css');
 @import url('../features/companies/css/styles.css');
