+++ Frontend/src/utilities/hooks/useAuthContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '../context';
+import { AuthContext } from '../context/auth/authContext';
 
 export function useAuthContext() {
   return useContext(AuthContext);
